refactor: Controler의 검증 로직을 CarRace에 분리
