+++ src/View/InputView.js
@@ -1,9 +1,8 @@
 import readLine from 'readline';
 
 const INPUT_MESSAGES = Object.freeze({
-  carMessage:
-    '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분). ',
-  tryMessage: '시도할 횟수는 몇 회인가요?',
+  CAR_MESSAGE: '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분). ',
+  TRY_MESSAGE: '시도할 횟수는 몇 회인가요?',
 });
 
 const DIVIDE_SYMBOL = ',';
@@ -21,7 +20,6 @@ const InputView = {
 
   async readTry() {
     const tryNums = await this.readLineAsync(INPUT_MESSAGES.TRY_MESSAGE);
-
     return Number(tryNums);
   },
